Move booleanizer and MLTL instructions to separate files to mimic Rust version
